+++ druid/src/data.rs
@@ -415,6 +415,12 @@ impl Data for piet::FontStyle {
     }
 }
 
+impl Data for piet::TextAlignment {
+    fn same(&self, other: &Self) -> bool {
+        self == other
+    }
+}
+
 #[cfg(feature = "im")]
 impl<T: Data> Data for im::Vector<T> {
     fn same(&self, other: &Self) -> bool {